Update to OLED on Model-B
